処理の切り離し
