+++ app/_utils/api/apiUrl.ts
@@ -29,6 +29,8 @@ export const SHOPS_URL = {
   GET_WITH_REVIEWS: (shopId: number) => `/shops/${shopId}/reviews`,
   // 作成
   POST: '/shops',
+  // 編集
+  PUT: (shopId: number) => `/shops/${shopId}`,
   // 削除
   DELETE: (shopId: number) => `/shops/${shopId}`
 }
